+++ TravelSchedule/Views/Carriers/CarriersListView.swift
@@ -25,9 +25,14 @@ struct CarriersListView: View {
                     Spacer()
                 } else {
                     List(viewModel.filteredRoutes) { route in
-                        CarriersRowView(route: route)
-                            .listRowInsets(EdgeInsets(top: 4, leading: 9, bottom: 4, trailing: 8))
-                            .listRowBackground(Color.clear)
+                        Button(action: {
+                            navigationPath.append(Destination.carrierDetail(route: route))
+                        }) {
+                            CarriersRowView(route: route)
+                                .listRowInsets(EdgeInsets(top: 4, leading: 9, bottom: 4, trailing: 8))
+                                .listRowBackground(Color.clear)
+                        }
+                        .buttonStyle(PlainButtonStyle())
                     }
                     .listStyle(.plain)
                     .background(Color.clear)
@@ -45,7 +50,7 @@ struct CarriersListView: View {
             VStack {
                 Spacer()
                 Button(action: {
-                    navigationPath.append(ContentView.Destination.filters(
+                    navigationPath.append(Destination.filters(
                         fromCity: fromCity,
                         fromStation: fromStation,
                         toCity: toCity,
@@ -72,7 +77,6 @@ struct CarriersListView: View {
         }
     }
 }
-
 
 #Preview {
     CarriersListView(
