+++ TravelSchedule/Views/Cities/CitiesView.swift
@@ -53,7 +53,7 @@ struct CitiesView: View {
                         ForEach(filteredCities) { city in
                             Button(action: {
                                 selectedCity = city
-                                navigationPath.append(ContentView.Destination.stations(city: city, isSelectingFrom: isSelectingFrom))
+                                navigationPath.append(Destination.stations(city: city, isSelectingFrom: isSelectingFrom))
                             }) {
                                 CityRowView(city: city)
                                     .listRowInsets(EdgeInsets(top: 4, leading: 9, bottom: 4, trailing: 4))
